requirements.txt
et-xmlfile==1.1.0
Django==3.2.11
django-cors-headers==3.13.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2 
pymongo==3.12.1
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.2.4
PyPDF2==2.1.0

asgiref==3.5.2
distlib==0.3.4
dj-database-url==0.5.0
django-heroku==0.3.1
et-xmlfile==1.1.0
filelock==3.7.1
gunicorn==20.1.0
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2022.1
tk==0.1.0
virtualenv==20.14.1
whitenoise==6.2.0